Windows port: experimental isin: fix MSVC narrowing conversions
- isin_device_operation.cpp, isin_program_factory.cpp: add static_cast<uint32_t>
  for uint64_t logical_volume() values placed into a Shape{} braced initializer
  and a std::vector<uint32_t> compile-time-args list, which MSVC treats as
  narrowing.

+++ ttnn/cpp/ttnn/operations/experimental/isin/device/isin_program_factory.cpp
@@ -89,8 +89,8 @@ IsInProgramFactory::cached_program_t IsInProgramFactory::create(
         static_cast<uint32_t>(IsInCB::ELEMENTS),
         static_cast<uint32_t>(IsInCB::TEST_ELEMENTS),
         static_cast<uint32_t>(IsInCB::OUTPUT),
-        elements_tensor.logical_volume(),
-        test_elements_tensor.logical_volume(),
+        static_cast<uint32_t>(elements_tensor.logical_volume()),
+        static_cast<uint32_t>(test_elements_tensor.logical_volume()),
         single_fetch_subchunk_size,
         static_cast<uint32_t>(invert),
         elements_tensor.element_size()};
